+++ amiadapters/base.py
