+++ amiadapters/config.py
